feat: se elimina código comentado en componente de listar-plan
