+++ src/app/pages/plan/visualizar-fechas/visualizar-fechas.component.scss
@@ -42,10 +42,6 @@ th {
   color: white;
 }
 
-td {
-  // white-space: nowrap; /* Evita que el texto se envuelva en pantallas pequeñas */
-}
-
 /* Hacer que la tabla tenga scroll horizontal en pantallas pequeñas */
 .table-responsive {
   width: 100%;
